feat(core): add reset option to firestore
BREAKING CHANGE: When unbinding a collection or document in Firestore,
the data is, by default reset to `null` for documents or an empty
array `[]` for collections. This option can be set to `false` to keep
last value instead or to a function returning a value to customize the
new value.